add: implement UART and basic SBI

+++ os/src/console.rs
@@ -1,5 +1,6 @@
 // os/src/console.rs
-use crate::sbi::console_putchar;
+use crate::sbi::console_putchar; 
+// use crate::sbi::console_putchar;
 use core::fmt::{self, Write};
 
 struct Stdout;
@@ -7,7 +8,7 @@ struct Stdout;
 impl Write for Stdout {
     fn write_str(&mut self, s: &str) -> fmt::Result {
         for c in s.chars() {
-            console_putchar(c as usize);
+            console_putchar(c as u8);
         }
         Ok(())
     }
